MNT #525 avoid circular import
