+++ apstools/plans.py
@@ -4,6 +4,7 @@ Plans that might be useful at the APS when using Bluesky
 .. autosummary::
 
    ~addDeviceDataAsStream
+   ~documentation_run
    ~execute_command_list
    ~get_command_list
    ~lineup
@@ -64,35 +65,6 @@ class CommandFileReadError(IOError):
 
     .. index:: Bluesky Exception; CommandFileReadError
     """
-
-
-def addDeviceDataAsStream(devices, label):
-    """
-    plan: add an ophyd Device as an additional document stream
-
-    .. index:: Bluesky Plan; addDeviceDataAsStream
-
-    Use this within a custom plan, such as this example::
-
-        from apstools.plans import addDeviceStream
-        ...
-        yield from bps.open_run()
-        # ...
-        yield from addDeviceDataAsStream(prescanDeviceList, "metadata_prescan")
-        # ...
-        yield from custom_scan_procedure()
-        # ...
-        yield from addDeviceDataAsStream(postscanDeviceList, "metadata_postscan")
-        # ...
-        yield from bps.close_run()
-
-    """
-    yield from bps.create(name=label)
-    if not isinstance(devices, list):  # just in case...
-        devices = [devices]
-    for d in devices:
-        yield from bps.read(d)
-    yield from bps.save()
 
 
 def execute_command_list(filename, commands, md=None):
